Рефакторинг кода

+++ src/main/java/dev/radagon/trello/controller/BoardColumnController.java
@@ -1,5 +1,6 @@
 package dev.radagon.trello.controller;
 
+import dev.radagon.trello.config.SecurityHelper;
 import dev.radagon.trello.dto.BoardColumnDTO;
 import dev.radagon.trello.entity.Board;
 import dev.radagon.trello.entity.BoardColumn;
@@ -7,13 +8,9 @@ import dev.radagon.trello.entity.User;
 import dev.radagon.trello.service.BoardColumnService;
 import dev.radagon.trello.service.BoardService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.*;
-
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.web.server.ResponseStatusException;
 
 @Controller
 @RequiredArgsConstructor
@@ -22,28 +19,22 @@ public class BoardColumnController {
     private final BoardColumnService columnService;
     private final BoardService boardService;
 
+    private final SecurityHelper securityHelper;
+
     @PostMapping("/boards/{boardId}/columns")
     public String createColumn(
             @PathVariable String publicId,
             @PathVariable Long boardId,
             @ModelAttribute BoardColumnDTO dto,
-            Authentication auth
+            Authentication authentication
             ) {
 
-        String currentEmail = auth.getName();
+        User currentUser = securityHelper.getAuthenticatedUser(publicId, authentication);
+        if (currentUser == null) {return  "redirect:/login";}
 
-//        User owner = boardService.getOwnerByPublicId(publicId);
-        User currentUser = boardService.getUserByEmail(currentEmail);
-
-
-        try {
-            columnService.createColumn(dto, boardId, currentUser.getId());
-        } catch (AccessDeniedException e) {
-            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "Нет прав на эту доску");
-        }
+        columnService.createColumn(dto, boardId, currentUser.getId());
 
         Board board = boardService.getBoardById(boardId);
-
         return "redirect:/" + publicId + "/" + board.getSlug();
     }
 
@@ -52,29 +43,25 @@ public class BoardColumnController {
             @PathVariable String publicId,
             @PathVariable Long columnId,
             @ModelAttribute BoardColumnDTO dto,
-            Authentication auth) {
+            Authentication authentication) {
 
-        String currentEmail = auth.getName();
-        User currentUser = boardService.getUserByEmail(currentEmail);
+        User currentUser = securityHelper.getAuthenticatedUser(publicId, authentication);
+        if (currentUser == null) {return  "redirect:/login";}
 
         BoardColumn column = columnService.updateColumn(columnId, dto, currentUser.getId());
-
-        Board board = column.getBoard();
-
-        return "redirect:/" + publicId + "/" + board.getSlug();
+        return "redirect:/" + publicId + "/" + column.getBoard().getSlug();
     }
 
     @PostMapping("/columns/{columnId}/delete")
     public String deleteColumn(
             @PathVariable String publicId,
             @PathVariable Long columnId,
-            Authentication auth) {
+            Authentication authentication) {
 
-        String currentEmail = auth.getName();
-        User currentUser = boardService.getUserByEmail(currentEmail);
+        User currentUser = securityHelper.getAuthenticatedUser(publicId, authentication);
+        if (currentUser == null) {return  "redirect:/login";}
 
         Board board = columnService.deleteColumn(columnId,  currentUser.getId());
-
         return "redirect:/" + publicId + "/" + board.getSlug();
     }
 }
